feat(api): add pundit policies and enforce user-scoped authorization

+++ api/app/controllers/notes_controller.rb
@@ -2,8 +2,10 @@ class NotesController < ApplicationController
   before_action :authenticate_user!
 
   def create
-    application = current_user.applications.find(params[:application_id])
+    application = policy_scope(Application).find(params[:application_id])
+    authorize application, :show?
     note = application.notes.new(note_params)
+    authorize note
 
     if note.save
       render json: { data: note.as_json(only: %i[id application_id body created_at updated_at]) }, status: :created
@@ -13,7 +15,8 @@ class NotesController < ApplicationController
   end
 
   def destroy
-    note = Note.joins(:application).where(applications: { user_id: current_user.id }).find(params[:id])
+    note = policy_scope(Note).find(params[:id])
+    authorize note
     note.destroy!
     head :no_content
   end
